Use haskell.nix's built-in tools support for `shellFor`.
haskell.nix has a special mechanism for making tools (i.e., Haskell
executables) available in a `shellFor`, so we take advantage of that
where we can. Note that brittany doesn't currently work with this
mechanism, and hie and hls aren't in Hackage yet, so we still handle
these as special-case buildInputs. Ditto for cabal-fmt, which has new
functionality that we want to use that hasn't been pushed to Hackage
yet.