cabalProject: `ghc` is now a required parameter.
When I applied the previous refactoring commits, I lost the bit of the
old `cabalProject883` function that overrides alex for a GHC 8.8.3
shell, breaking it in the process. This commit restores that
functionality, but does so in a way that requires that the caller
provide an explicit GHC version. This seems like good practice,
anyway.