Materialize (most) everything in our shells.
This should hopefully improve performance, both for the time it takes
from running `nix-shell` until the prompt appears; and for Hydra
evaluations of haskell-hacknix-based projects.

Note that this process to be re-run whenever we bump package versions
etc. Right now, it needs to be done by hand (and per-shell), but we
can probably write a script that automates it.

Also: upstream doesn't materialize hoogle, so that still results in a
bit of slowness. We should probably override it ourselves; see #69.