Remove collectChecks function.
There's an upstream version that we can use instead. Note that the
upstream version's filter function only takes the *package* as an
argument, not the `name: package:` pair of arguments, which means that
it's not quite as easy to filter by package name: haskell.nix Haskell
packages are supposed to have an `identifier.name` attribute, but the
Win32 package, if not present, causes this filter to break; see:

https://github.com/input-output-hk/haskell.nix/blob/c88f9eccc975b21ae1e6a6b8057a712b91e374f2/lib/default.nix#L132

However, we can use (and probably should be using) haskell.nix's
`isLocal` filter anyway, rather than a fragile package name/prefix
filter. I think that will suffice, and therefore we can get rid of our
own collectChecks function.