Stop trying to materialize haskell-hacknix packages.
Because:

1. These are a pain to maintain and update. The scripts don't really
work (because their materialization directories point directly into
the Nix store) and need to be run by hand, and are even a bigger pain
to run cross-platform.

2. The materialization speedups are pretty minimal compared to the
cost of building the packages themselves.

3. The materialization speedups are also meant to make evaluations
faster, but a Hydra needs to check them anyway (and therefore doesn't
benefit from them), and Nix Flakes' `nix develop` means that evals can
be cached and therefore materialization has no impact on the typical
development workflow, either.