Bump dependencies and update index.state to latest.
Also, we need some fixes, as things we were doing previously no longer
work as of this upstream commit:

https://github.com/input-output-hk/haskell.nix/pull/1087/files#

Apparently the `subdir` argument to cabalProject was never supported.
We replace the old subdir arg with haskell-nix's cleanGit. See
https://input-output-hk.github.io/haskell.nix/tutorials/clean-git/

Our extra-custom-tools override no longer works, so we just build
cabal-fmt directly and include it in the shellFor buildInputs.